Move orignal main to folder Samples. Add utils class for getting token easily. Refactor OrderTests for use this Utils.
